requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
idna==2.8
Jinja2==2.10.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
urllib3==1.25.7
